+++ src/lib/components/event/event-details-row-expanded.svelte
@@ -15,7 +15,7 @@
   } from '$lib/utilities/get-single-attribute-for-event';
 
   import CodeBlock from '$lib/holocene/code-block.svelte';
-  import Link from '../link.svelte';
+  import Link from '$lib/holocene/link.svelte';
   import Copyable from '../copyable.svelte';
 
   export let key: string;
