+++ src/lib/components/workflow/pending-activities.svelte
@@ -8,7 +8,7 @@
     formatDuration,
   } from '$lib/utilities/format-date';
   import { routeForPendingActivities } from '$lib/utilities/route-for';
-  import Link from '$lib/components/link.svelte';
+  import Link from '$lib/holocene/link.svelte';
   import Icon from '$holocene/icon/index.svelte';
   import Badge from '$holocene/badge.svelte';
   import CodeBlock from '$lib/holocene/code-block.svelte';
